+++ src/main/java/com/idetix/identityapprover/service/email/EmailServiceImpl.java
@@ -37,5 +37,4 @@ public class EmailServiceImpl implements EmailService {
         return result;
     }
 
-
 }
