+++ src/main/java/com/idetix/identityapprover/service/email/EmailServiceMockImpl.java
@@ -12,5 +12,4 @@ public class EmailServiceMockImpl implements EmailService {
         return true;
     }
 
-
 }
