+++ src/main/java/com/idetix/identityapprover/service/phone/PhoneIdentityService.java
@@ -60,7 +60,6 @@ public class PhoneIdentityService {
                 updatePhoneIdentity(phoneIdentity);
             }
         }
-
         return phoneIdentity;
     }
