+++ src/test/java/com/idetix/identityapprover/service/security/SecurityServiceImplTest.java
@@ -1,7 +1,6 @@
-package com.idetix.identityapprover.service;
+package com.idetix.identityapprover.service.security;
 
-import com.idetix.identityapprover.service.security.SecurityService;
-import com.idetix.identityapprover.service.security.SecurityServiceImpl;
+
 import org.junit.jupiter.api.Test;
 
 import java.text.CharacterIterator;
